# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Parmana API
  version: 1.0.0
  description: |
    Parmana sits between an AI agent and the systems it acts on. An agent sends a Business Transaction: who
    authorized it, what action it wants, and the facts that justify it. Parmana evaluates the policy bound to
    the action, checks any signed human approval the policy requires, releases the action to the connector
    only if the policy approves, and returns a signed Execution Trust Record anyone can verify.

    **Authentication.** Send `Authorization: Bearer <key>` on every request. The only routes that need no key
    are the probes, the documents and the independent verification routes: `GET /health`, `GET /ready`,
    `GET /openapi.yaml`, `GET /openapi.json`, `GET /api-manifest.json`, `GET /documentation`,
    `GET /reference`, `GET /parmana-handbook.pdf`, `GET /handbook/download-leads`,
    `POST /handbook/download-leads`, `POST /refusal/verify`, `POST /execution-intents/verify`,
    `POST /audit/verify`, `GET /keys/{keyId}` and `GET /.well-known/jwks.json`. A missing or unknown key
    returns `401` before anything else runs. The server holds only a hash of each key. See
    [Authentication](/api-reference/authentication).

    **Two kinds of key.** An agent or service key may invoke only the capabilities listed on it. A human key
    (`credentialHolderType: USER`) is required for governance: proposing and approving policies, approvers
...
    person who proposed it.

...

...

...

  contact:
...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /transactions/{businessTransactionId}:
    get:
      tags:
        - Transactions
      summary: Get Business Transaction
      operationId: getTransaction
      description: |
        Returns one Business Transaction as it was stored. For the signed outcome, what was decided and
        executed, use `GET /trust-records/{businessTransactionId}`.
      parameters:
        - name: businessTransactionId
          in: path
          required: true
          description: "The id you chose when you sent the transaction."
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: "The Business Transaction."
          content:
            application/json:
              schema:
                $ref: "../schemas/responses/transaction-response.schema.json"
              examples:
                vendor-payment-flow-transactions-endpoint:
                  summary: Real captured response
                  value:
                    businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                    metadata:
                      businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                      correlationId: cc1ca975-4935-4e9a-a591-8897af6b56fb
                      sourceSystem: vendor-payment-service
                      submittedBy: ap-automation
                      submittedAt: "2026-07-13T17:36:00.835Z"
                    authority:
                      authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                      authorityType: SERVICE
                      principalId: ap-automation-svc
                      displayName: Accounts Payable Automation
                      issuedAt: "2026-07-13T17:36:00.835Z"
                    authorization:
                      authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                      authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                      purpose: Authorize vendor payment disbursement
                      issuedAt: "2026-07-13T17:36:00.835Z"
                    intent:
                      intentId: 1e77d0b3-2e57-4e31-b780-03abc92f1b69
                      authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                      action: payments:execute
                      target: vendor/V-500
                      parameters:
                        amount: 5000
                        currency: USD
                      createdAt: "2026-07-13T17:36:00.835Z"
                    policy:
                      name: vendor-payment
                      version: 2.0.0
                      schemaVersion: 1.0.0
                    signals:
                      vendorVerified: true
                      invoiceVerified: true
                      paymentApproved: true
                      sufficientFunds: true
                      paymentAmount: 5000
                      riskScore: 12
                    status: RECEIVED
                    createdAt: "2026-07-13T17:36:00.965Z"

        "404":
          description: "Your key submitted no transaction with this id."
          content:
            application/json:
              schema:
                $ref: "../schemas/common/error.schema.json"
              examples:
                notFound:
                  summary: Real captured response
                  value:
                    error: Business Transaction not found.
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      # Prefills the docs playground with the public sandbox's demo key
      # (ADR-0014): caller sandbox-visitor, allowed only sandbox:receipt.
      x-default: 2VfYWCzt_cBAPK-8uufX6ordfY2JuQhFPsohuEumKME
      description: >
        Caller API key issued by scripts/generate-api-key.ts. Sent as
        Authorization: Bearer <key>. Verified against a stored SHA-256 hash
        in constant time by packages/api/src/auth/StaticKeyAuthenticator.ts.
        Required on every route not listed as exempt in this document's
        top-level description. See /api-reference/authentication.
  responses:
    Unauthorized:
      description: "The `Authorization` header is missing, is not a bearer key, or names a key the server does not know. Nothing ran. Send a valid key."
      content:
        application/json:
          schema:
            $ref: "../schemas/common/error.schema.json"
          examples:
            authRequired:
              summary: Real captured response, missing or invalid Authorization header
              value:
                error: authentication required
...